test update

+++ python/cudf/cudf/tests/test_multiindex.py
@@ -753,7 +753,7 @@ def test_multiindex_copy_sem(data, levels, codes, names):
     for glv, plv in zip(gmi_copy.levels, pmi_copy.levels):
         assert all(glv.values_host == plv.values)
     for gval, pval in zip(gmi.codes, pmi.codes):
-        assert_eq(gval, pval.astype(np.int64))
+        assert_eq(gval, pval)
     assert_eq(gmi_copy.names, pmi_copy.names)
 
     # Test same behavior when used on DataFrame
